refactor(CalendarDropdownMenu): streamline component structure and enhance styling; remove redundant file
