+++ components/widgets/Calendar/DropdownMenu/Combobox.tsx
@@ -12,11 +12,7 @@ import {
   CommandItem,
   CommandList,
 } from '@/components/ui/command';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/popover';
+import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 
 type ComboboxProps = {
   options: number[];
@@ -24,11 +20,7 @@ type ComboboxProps = {
   onChangeVal: (newValue: number) => void;
 };
 
-export default function Combobox({
-  options,
-  value,
-  onChangeVal,
-}: ComboboxProps) {
+export default function Combobox({ options, value, onChangeVal }: ComboboxProps) {
   const [open, setOpen] = useState(false);
 
   return (
@@ -38,13 +30,13 @@ export default function Combobox({
           variant="outline"
           role="combobox"
           aria-expanded={open}
-          className="h-8 w-16 justify-between px-2 text-xs"
+          className="h-8 w-20 justify-between px-2 text-sm"
         >
           {value}
           <ChevronsUpDown className="ml-2 h-3 w-3 shrink-0 opacity-50" />
         </Button>
       </PopoverTrigger>
-      <PopoverContent className="w-16 p-0">
+      <PopoverContent className="w-28 p-0">
         <Command>
           <CommandInput placeholder="Search..." />
           <CommandList>
@@ -60,10 +52,7 @@ export default function Combobox({
                   }}
                 >
                   <Check
-                    className={cn(
-                      'mr-2 h-4 w-4',
-                      value === option ? 'opacity-100' : 'opacity-0',
-                    )}
+                    className={cn('mr-2 h-4 w-4', value === option ? 'opacity-100' : 'opacity-0')}
                   />
                   {option}
                 </CommandItem>
